Fix warnings
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2065 111d7451-93fa-4e81-90d4-8039f76616e9

+++ dictobj.h
@@ -41,6 +41,6 @@ struct pair_s {
     struct avltree_node node;
 };
 
-extern Obj *dictobj_parse(struct values_s *, unsigned int);
+extern Obj *dictobj_parse(struct values_s *, size_t);
 
 #endif
